refactor: 가수 디테일과 검색결과 type 분리
