+++ frontend/src/features/search/types/search.type.ts
@@ -1,20 +1,5 @@
-interface SingersSong {
-  id: number;
-  title: string;
-  albumCoverUrl: string;
-  videoLength: number;
-}
-
 export interface SingerSearchPreview {
   id: number;
   singer: string;
   profileImageUrl: string;
-}
-
-export interface SingerSearchResult {
-  id: number;
-  singer: string;
-  profileImageUrl: string;
-  totalSongCount: number;
-  songs: SingersSong[];
 }
